fix(domain/data): resolve compilation errors and add unit tests for 60% coverage
Fixes import placement, override keywords, and type mismatches. Adds unit tests for domain and data layers.

+++ domain/src/main/java/app/otakureader/domain/repository/ReaderSettingsRepository.kt
@@ -21,7 +21,7 @@ interface ReaderSettingsRepository {
     val fullscreen: Flow<Boolean>
     val incognitoMode: Flow<Boolean>
     val colorFilterMode: Flow<ColorFilterMode>
-    val customTintColor: Flow<Int>
+    val customTintColor: Flow<Long>
     val cropBordersEnabled: Flow<Boolean>
     val imageQuality: Flow<ImageQuality>
     val dataSaverEnabled: Flow<Boolean>
